CI: a cast gcc calls useless and a comparison clang-tidy calls std::max
`gaps.size()` is already `size_t`, so casting it is `-Wuseless-cast` at
every gcc variant. And `order_tests.cpp` never included `scav_int.h`, so
the loop that was reaching for `imax` had written the comparison out by
hand; the include makes it `imax` again, which is both what the rest of
the tree does and not a `>` for readability-use-std-min-max to find.

+++ src/layout/order.cpp
@@ -550,7 +550,7 @@ SubmachineOrders order_submachines(Chart const &c,
       if (dense[root] == INVALID) { dense[root] = parts++; }
     }
     std::vector<uint32_t> &lanes{ sc.lanes };
-    lanes.assign(static_cast<size_t>(gaps.size()) * parts, 0);
+    lanes.assign(gaps.size() * parts, 0);
     int32_t const pitch{ label_line_height(p) };
     auto const turn = [&](uint32_t b, uint32_t of) {
       if (b >= gaps.size()) { return; }
